+++ tests/test_data/test_config.py
@@ -1,6 +1,8 @@
 def get_test_config() -> dict:
     return {
         'mfa_shell_command': 'some-command',
+        'shell_path_extension': '/some/dir/',
         'default_access_key': 'some-access-key',
         'default_sso_session': 'some-sso-session',
+        'default_sso_interval': 'some-sso-interval',
     }